# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully."
          },
          "409": {
            "description": "Email already exists."
          }
        },
        "summary": "Register a customer account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful and access token returned."
          },
          "401": {
            "description": "Invalid email or password."
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Refresh token rotated successfully."
          },
          "401": {
            "description": "Invalid refresh token."
          }
        },
        "summary": "Rotate refresh token and issue a new access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully."
          }
        },
        "summary": "Request password reset OTP (Mock)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully."
          },
          "401": {
            "description": "Invalid active session."
          }
        },
        "summary": "Logout current device session",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoWayHome API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "phone": {
            "type": "string",
            "example": "+84901234567"
          },
          "userType": {
            "type": "string",
            "enum": [
              "customer",
              "partner",
              "admin"
            ],
            "example": "customer"
          },
          "businessName": {
            "type": "string",
            "example": "Nguyen Van A Travel Co."
          },
          "businessType": {
            "type": "string",
            "enum": [
              "individual",
              "company"
            ],
            "example": "individual"
          }
        },
        "required": [
          "email",
          "password",
          "fullName",
          "phone",
          "userType"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123",
            "minLength": 8
          },
          "deviceName": {
            "type": "string",
            "example": "Chrome on Windows"
          },
          "deviceType": {
            "type": "string",
            "enum": [
              "web",
              "ios",
              "android",
              "other"
            ],
            "example": "web"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          }
        },
        "required": [
          "email"
        ]
      },
...
    }
  }
}